requirements.txt
python-telegram-bot
orator
redis